+++ applications/ChatGPT/chatgpt/models/base/__init__.py
@@ -1,0 +1,5 @@
+from .actor import Actor
+from .critic import Critic
+from .reward_model import RewardModel
+
+__all__ = ['Actor', 'Critic', 'RewardModel']
