+++ applications/ChatGPT/chatgpt/models/bloom/bloom_critic.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 from transformers import BloomConfig, BloomForCausalLM, BloomModel
 
-from .critic import Critic
+from ..base import Critic
 
 
 class BLOOMCritic(Critic):
