+++ applications/ChatGPT/chatgpt/models/gpt/__init__.py
@@ -1,0 +1,5 @@
+from .gpt_actor import GPTActor
+from .gpt_critic import GPTCritic
+from .gpt_rm import GPTRM
+
+__all__ = ['GPTActor', 'GPTCritic', 'GPTRM']
