+++ applications/ChatGPT/chatgpt/models/gpt/gpt_rm.py
@@ -1,36 +1,38 @@
 from typing import Optional
 
 import torch.nn as nn
-from transformers import BloomConfig, BloomForCausalLM, BloomModel
+from transformers.models.gpt2.configuration_gpt2 import GPT2Config
+from transformers.models.gpt2.modeling_gpt2 import GPT2Model
 
-from .reward_model import RewardModel
+from ..base import RewardModel
 
 
-class BLOOMRM(RewardModel):
+class GPTRM(RewardModel):
     """
-    BLOOM Reward model.
+    GPT Reward model.
 
     Args:
         pretrained (str): Pretrained model name or path.
-        config (BloomConfig): Model config.
+        config (GPT2Config): Model config.
         checkpoint (bool): Enable gradient checkpointing.
-        lora_rank (int): LoRA rank.
+        lora_rank (int): Rank of the low-rank approximation.
         lora_train_bias (str): LoRA bias training mode.
     """
 
     def __init__(self,
-                 pretrained: str = None,
-                 config: Optional[BloomConfig] = None,
+                 pretrained: Optional[str] = None,
+                 config: Optional[GPT2Config] = None,
                  checkpoint: bool = False,
                  lora_rank: int = 0,
                  lora_train_bias: str = 'none') -> None:
         if pretrained is not None:
-            model = BloomModel.from_pretrained(pretrained)
+            model = GPT2Model.from_pretrained(pretrained)
         elif config is not None:
-            model = BloomModel(config)
+            model = GPT2Model(config)
         else:
-            model = BloomModel(BloomConfig())
+            model = GPT2Model(GPT2Config())
         if checkpoint:
             model.gradient_checkpointing_enable()
-        value_head = nn.Linear(model.config.hidden_size, 1)
+
+        value_head = nn.Linear(model.config.n_embd, 1)
         super().__init__(model, value_head, lora_rank, lora_train_bias)
