+++ applications/ChatGPT/chatgpt/models/opt/__init__.py
@@ -1,0 +1,5 @@
+from .opt_actor import OPTActor
+from .opt_critic import OPTCritic
+from .opt_rm import OPTRM
+
+__all__ = ['OPTActor', 'OPTCritic', 'OPTRM']
