+++ applications/ChatGPT/chatgpt/models/opt/opt_actor.py
@@ -1,19 +1,18 @@
 from typing import Optional
 
-import torch.nn as nn
-from transformers.models.gpt2.configuration_gpt2 import GPT2Config
-from transformers.models.gpt2.modeling_gpt2 import GPT2Model
+from transformers.models.opt.configuration_opt import OPTConfig
+from transformers.models.opt.modeling_opt import OPTForCausalLM
 
-from .reward_model import RewardModel
+from ..base import Actor
 
 
-class GPTRM(RewardModel):
+class OPTActor(Actor):
     """
-    GPT Reward model.
+    OPT Actor model.
 
     Args:
         pretrained (str): Pretrained model name or path.
-        config (GPT2Config): Model config.
+        config (OPTConfig): Model config.
         checkpoint (bool): Enable gradient checkpointing.
         lora_rank (int): Rank of the low-rank approximation.
         lora_train_bias (str): LoRA bias training mode.
@@ -21,18 +20,16 @@ class GPTRM(RewardModel):
 
     def __init__(self,
                  pretrained: Optional[str] = None,
-                 config: Optional[GPT2Config] = None,
+                 config: Optional[OPTConfig] = None,
                  checkpoint: bool = False,
                  lora_rank: int = 0,
                  lora_train_bias: str = 'none') -> None:
         if pretrained is not None:
-            model = GPT2Model.from_pretrained(pretrained)
+            model = OPTForCausalLM.from_pretrained(pretrained)
         elif config is not None:
-            model = GPT2Model(config)
+            model = OPTForCausalLM(config)
         else:
-            model = GPT2Model(GPT2Config())
+            model = OPTForCausalLM(OPTConfig())
         if checkpoint:
             model.gradient_checkpointing_enable()
-
-        value_head = nn.Linear(model.config.n_embd, 1)
-        super().__init__(model, value_head, lora_rank, lora_train_bias)
+        super().__init__(model, lora_rank, lora_train_bias)
