+++ applications/ChatGPT/chatgpt/trainer/strategies/base.py
@@ -5,7 +5,7 @@ from typing import Any, List, Tuple, Union
 import numpy as np
 import torch
 import torch.nn as nn
-from chatgpt.nn import Actor
+from chatgpt.models.base import Actor, Critic, RewardModel
 from chatgpt.replay_buffer import ReplayBuffer
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader
